Exception handling, bug fixing

+++ src/main/java/com/technomarket/technomarket/service/impl/ProductServiceImpl.java
@@ -1,20 +1,19 @@
 package com.technomarket.technomarket.service.impl;
 
-import com.technomarket.technomarket.entity.Feedback;
+import com.technomarket.technomarket.entity.Review;
 import com.technomarket.technomarket.entity.Product;
 import com.technomarket.technomarket.entity.User;
-import com.technomarket.technomarket.entity.enums.Status;
-import com.technomarket.technomarket.repository.FeedbackRepository;
+import com.technomarket.technomarket.repository.ReviewRepository;
 import com.technomarket.technomarket.repository.ProductRepository;
 import com.technomarket.technomarket.repository.UserRepository;
 import com.technomarket.technomarket.service.ProductService;
+import com.technomarket.technomarket.service.impl.exceptions.ResourceNotFoundException;
+import com.technomarket.technomarket.service.impl.exceptions.UnauthorizedProductAccessException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.security.Principal;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -22,13 +21,13 @@ import java.util.List;
 public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
     private final UserRepository userRepository;
-    private final FeedbackRepository feedbackRepository;
+    private final ReviewRepository reviewRepository;
 
     @Autowired
-    public ProductServiceImpl(ProductRepository productRepository, UserRepository userRepository, FeedbackRepository feedbackRepository) {
+    public ProductServiceImpl(ProductRepository productRepository, UserRepository userRepository, ReviewRepository reviewRepository) {
         this.productRepository = productRepository;
         this.userRepository = userRepository;
-        this.feedbackRepository = feedbackRepository;
+        this.reviewRepository = reviewRepository;
     }
 
     @Override
@@ -40,13 +39,25 @@ public class ProductServiceImpl implements ProductService {
     }
 
     public User getUserByPrincipal(Principal principal) {
-        if (principal == null) return new User();
+        if (principal == null) {
+            throw new UnauthorizedProductAccessException("Unauthorized request, please login first");
+        }
         return userRepository.findByUsername(principal.getName());
     }
 
     @Override
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public void deleteProduct(Long id, Principal principal) {
+        Product product = productRepository.findById(id).orElse(null);
+        if (product == null){
+            throw new ResourceNotFoundException("Product not found with id: " + id);
+        }
+        User owner  = product.getUser();
+        if (owner.getUsername().equals(principal.getName())){
+            productRepository.deleteById(id);
+            log.info("Product with id {} was deleted by {}", id, owner);
+        } else {
+            throw new UnauthorizedProductAccessException("You don't have enough authority to delete this product");
+        }
     }
 
     @Override
@@ -56,15 +67,21 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public Product getProductById(Long id) {
-        return productRepository.findById(id).orElse(null);
+        Product product = productRepository.findById(id).orElse(null);
+        if (product == null){
+            throw new ResourceNotFoundException("Product not found with id: " + id);
+        } else {
+            return product;
+        }
+
     }
 
     @Override
-    public void createFeedback(Product product, Feedback feedback, Principal principal) {
-        feedback.setProduct(product);
-        feedback.setUser(getUserByPrincipal(principal));
-        log.info("Saving new feedback. Text: {}; Author username: {}", feedback.getText(), feedback.getUser().getUsername());
-        feedbackRepository.save(feedback);
+    public void createReview(Product product, Review review, Principal principal) {
+        review.setProduct(product);
+        review.setUser(getUserByPrincipal(principal));
+        log.info("Saving new feedback. Text: {}; Author username: {}", review.getText(), review.getUser().getUsername());
+        reviewRepository.save(review);
     }
